# server/go/model_new_conference.go
/*
 * Find a Conference
 *
 * API for Find a Conference's conference database.
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type NewConference struct {

	Name string `json:"name"`

	Theme string `json:"theme"`

	Description string `json:"description"`

	Url string `json:"url"`

	StartDate string `json:"startDate,omitempty"`

	EndDate string `json:"endDate,omitempty"`
}

// AssertNewConferenceRequired checks if the required fields are not zero-ed
func AssertNewConferenceRequired(obj NewConference) error {
	elements := map[string]interface{}{
		"name": obj.Name,
		"theme": obj.Theme,
		"description": obj.Description,
		"url": obj.Url,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertRecurseNewConferenceRequired recursively checks if required fields are not zero-ed in a nested slice.
// Accepts only nested slice of NewConference (e.g. [][]NewConference), otherwise ErrTypeAssertionError is thrown.
func AssertRecurseNewConferenceRequired(objSlice interface{}) error {
	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
		aNewConference, ok := obj.(NewConference)
		if !ok {
			return ErrTypeAssertionError
		}
		return AssertNewConferenceRequired(aNewConference)
	})
}